Rename org.pf4j.plugin package to org.pf4j.test
