+++ pf4j/src/test/java/org/pf4j/ManifestPluginDescriptorFinderTest.java
@@ -18,7 +18,7 @@ package org.pf4j;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
-import org.pf4j.plugin.PluginJar;
+import org.pf4j.test.PluginJar;
 
 import java.io.FileOutputStream;
 import java.io.IOException;
