+++ pf4j/src/test/java/org/pf4j/SingletonExtensionFactoryTest.java
@@ -16,8 +16,8 @@
 package org.pf4j;
 
 import org.junit.jupiter.api.Test;
-import org.pf4j.plugin.FailTestExtension;
-import org.pf4j.plugin.TestExtension;
+import org.pf4j.test.FailTestExtension;
+import org.pf4j.test.TestExtension;
 
 import java.io.File;
 import java.net.MalformedURLException;
