+++ pf4j/src/test/java/org/pf4j/test/AnotherFailTestPlugin.java
@@ -13,11 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.pf4j.plugin;
+package org.pf4j.test;
+
+import org.pf4j.Plugin;
 
 /**
  * @author Mario Franco
  */
-public class FailTestPlugin {
+public class AnotherFailTestPlugin extends Plugin {
+
+    public AnotherFailTestPlugin() {
+        super(null);
+    }
 
 }
