+++ pf4j/src/test/java/org/pf4j/test/TestPlugin.java
@@ -13,17 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.pf4j.plugin;
+package org.pf4j.test;
 
 import org.pf4j.Plugin;
+import org.pf4j.PluginWrapper;
 
 /**
  * @author Mario Franco
  */
-public class AnotherFailTestPlugin extends Plugin {
+public class TestPlugin extends Plugin {
 
-    public AnotherFailTestPlugin() {
-        super(null);
+    public TestPlugin(PluginWrapper wrapper) {
+        super(wrapper);
     }
 
 }
