+++ pf4j/src/test/java/org/pf4j/util/FileUtilsTest.java
@@ -17,7 +17,7 @@ package org.pf4j.util;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
-import org.pf4j.plugin.PluginZip;
+import org.pf4j.test.PluginZip;
 
 import java.nio.file.Files;
 import java.nio.file.Path;
